requirements.txt
streamlit~=1.17.0

Pillow~=9.4.0
numpy~=1.24.1